backend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
clarifai-grpc==9.9.0
click==8.1.7
cohere==4.27
colorama==0.4.6
comm==0.1.4
comtypes==1.4.1
contextlib2==21.6.0
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
Django==5.1.1
easyocr==1.7.1
executing==1.2.0
fastavro==1.8.2
fastjsonschema==2.18.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.53.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==3.0.0
grpcio==1.58.0
gTTS==2.5.1
h5py==3.11.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.35.1
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==3.3.2
kiwisolver==1.4.5
langchain==0.0.310
langsmith==0.0.43
lazy_loader==0.4
libclang==16.0.6
lxml==5.3.0
manifest-ml==0.0.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
ninja==1.11.1.1
nlpcloud==1.1.44
notebook_shim==0.2.3
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.84
openlm==0.0.5
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.20
pickleshare==0.7.5
pillow==10.3.0
platformdirs==3.10.0
playsound==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.2
pypiwin32==223
pytesseract==0.3.10
python-bidi==0.6.0
python-dateutil==2.8.2
python-json-logger==2.0.7
python-rapidjson==1.12
pyttsx3==2.90
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
redis==5.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.10.3
rsa==4.9
safetensors==0.4.0
schema==0.7.5
scikit-image==0.24.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.2
shapely==2.0.6
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.21
sqlitedict==2.1.0
sqlparse==0.5.1
stack-data==0.6.2
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.4.0
tifffile==2024.8.30
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.4.1
torchvision==0.19.1
tornado==6.3.3
tqdm==4.67.1
traitlets==5.10.0
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==2.3.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==14.2
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
